+++ lib/data/data.dart
@@ -1,75 +1,83 @@
-import 'package:flutter/material.dart';
-import 'package:instagram_ui_clone/model/models.dart'; 
+import 'package:instagram_ui_clone/model/models.dart';
 
 List<UserStory> getStories() {
   List<UserStory> userStories = new List();
   UserStory userStory = new UserStory();
 
   userStory.userName = "kadusuhas";
-  userStory.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userStory.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userStories.add(userStory);
 
   userStory = new UserStory();
   userStory.userName = "kadusuhas";
-  userStory.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userStory.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userStories.add(userStory);
 
   userStory = new UserStory();
   userStory.userName = "kadusuhas";
-  userStory.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userStory.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userStories.add(userStory);
 
-userStory = new UserStory();
+  userStory = new UserStory();
   userStory.userName = "kadusuhas";
-  userStory.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userStory.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userStories.add(userStory);
 
-userStory = new UserStory();
+  userStory = new UserStory();
   userStory.userName = "kadusuhas";
-  userStory.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userStory.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userStories.add(userStory);
 
-userStory = new UserStory();
+  userStory = new UserStory();
   userStory.userName = "kadusuhas";
-  userStory.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userStory.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userStories.add(userStory);
 
-userStory = new UserStory();
+  userStory = new UserStory();
   userStory.userName = "kadusuhas";
-  userStory.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userStory.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userStories.add(userStory);
 
-userStory = new UserStory();
+  userStory = new UserStory();
   userStory.userName = "kadusuhas";
-  userStory.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userStory.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userStories.add(userStory);
 
-userStory = new UserStory();
+  userStory = new UserStory();
   userStory.userName = "kadusuhas";
-  userStory.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userStory.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userStories.add(userStory);
 
-userStory = new UserStory();
+  userStory = new UserStory();
   userStory.userName = "kadusuhas";
-  userStory.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userStory.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userStories.add(userStory);
-
-
-
 
   return userStories;
 }
 
 List<UserPost> getPosts() {
-
   List<UserPost> userPosts = new List();
   UserPost userPost = new UserPost();
 
   userPost.userName = "kadusuhas100";
-  userPost.userAvatarImg = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userPost.userAvatarImg =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userPost.userSubtitle = "Pune, Maharashtra";
-  userPost.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
-  userPost.description = "Liked by suhask26 and others\nkadusuhas100 No need of caption!...more";
+  userPost.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userPost.description =
+      "Liked by suhask26 and others\nkadusuhas100 No need of caption!...more";
   userPost.comments = "View all comments\n30 minutes ago";
   userPosts.add(userPost);
 
@@ -83,24 +91,31 @@ List<UserPost> getPosts() {
 //   userPost = new UserPost();
 
   userPost.userName = "kadusuhas100";
-  userPost.userAvatarImg = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
-  userPost.userAvatarImg = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userPost.userAvatarImg =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userPost.userAvatarImg =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userPost.userSubtitle = "Pune, Maharashtra";
-  userPost.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
-  userPost.description = "Liked by suhask26 and others\nkadusuhas100 No need of caption!...more";
+  userPost.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userPost.description =
+      "Liked by suhask26 and others\nkadusuhas100 No need of caption!...more";
   userPost.comments = "View all comments\n30 minutes ago";
   userPosts.add(userPost);
   userPost = new UserPost();
-  
+
   userPost.userName = "kadusuhas100";
-  userPost.userAvatarImg = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userPost.userAvatarImg =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
   userPost.userSubtitle = "Pune, Maharashtra";
-  userPost.imgUrl = "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
-  userPost.description = "Liked by suhask26 and others\nkadusuhas100 No need of caption!...more";
+  userPost.imgUrl =
+      "https://avatars2.githubusercontent.com/u/60438083?s=460&u=8b777d70ae095c37b2efc63a5977fbe7f314f053&v=4";
+  userPost.description =
+      "Liked by suhask26 and others\nkadusuhas100 No need of caption!...more";
   userPost.comments = "View all comments\n30 minutes ago";
   userPosts.add(userPost);
-  
+
   userPost = new UserPost();
-   
+
   return userPosts;
 }
