+++ lib/views/home_page.dart
@@ -3,11 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:instagram_ui_clone/views/home_screen.dart';
 import 'package:instagram_ui_clone/views/notificationscreen.dart';
-import 'package:instagram_ui_clone/views/pages.dart';
 import 'package:instagram_ui_clone/views/post_screen.dart';
 import 'package:instagram_ui_clone/views/profile_screen.dart';
 import 'package:instagram_ui_clone/views/search_screen.dart';
-import 'package:instagram_ui_clone/widgets/widget.dart';
 
 class HomePage extends StatefulWidget {
   @override
@@ -17,20 +15,34 @@ class HomePage extends StatefulWidget {
 class _HomePageState extends State<HomePage> {
   int _selectedItem = 0;
 
-  List<Widget> pages = [HomeScreen(), SearchScreen(), PostScreen(), NotificaionScreen(), ProfileScreen()];
+  List<Widget> pages = [
+    HomeScreen(),
+    SearchScreen(),
+    PostScreen(),
+    NotificaionScreen(),
+    ProfileScreen()
+  ];
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(      
-      body: Center(child: pages[_selectedItem],),
+    return Scaffold(
+      body: Center(
+        child: pages[_selectedItem],
+      ),
       bottomNavigationBar: BottomNavigationBar(
         type: BottomNavigationBarType.fixed,
         elevation: 0.0,
         items: <BottomNavigationBarItem>[
           BottomNavigationBarItem(
-              icon: Icon(Icons.home_outlined, size: 30,),
+              icon: Icon(
+                Icons.home_outlined,
+                size: 30,
+              ),
               label: "",
-              activeIcon: Icon(Icons.home_filled, size: 30,)),
+              activeIcon: Icon(
+                Icons.home_filled,
+                size: 30,
+              )),
           BottomNavigationBarItem(
               icon: Icon(Icons.search),
               label: "",
@@ -42,9 +54,15 @@ class _HomePageState extends State<HomePage> {
               label: "",
               activeIcon: Icon(Icons.favorite)),
           BottomNavigationBarItem(
-              icon: FaIcon(FontAwesomeIcons.userCircle, size: 25,),
+              icon: FaIcon(
+                FontAwesomeIcons.userCircle,
+                size: 25,
+              ),
               label: "",
-              activeIcon: FaIcon(FontAwesomeIcons.solidUserCircle, size: 25,)),
+              activeIcon: FaIcon(
+                FontAwesomeIcons.solidUserCircle,
+                size: 25,
+              )),
         ],
         currentIndex: _selectedItem,
         onTap: (index) {
